QSPI EEPROM for SAMD51 (#17292)
