+++ Marlin/src/HAL/DUE/inc/Conditionals_post.h
@@ -21,7 +21,7 @@
  */
 #pragma once
 
-#if USE_EMULATED_EEPROM
+#if USE_FALLBACK_EEPROM
   #undef SRAM_EEPROM_EMULATION
   #undef SDCARD_EEPROM_EMULATION
   #define FLASH_EEPROM_EMULATION
