+++ Marlin/src/HAL/LINUX/eeprom_impl.cpp
@@ -26,7 +26,7 @@
 
 #if ENABLED(EEPROM_SETTINGS)
 
-#include "../shared/persistent_store_api.h"
+#include "../shared/eeprom_api.h"
 #include <stdio.h>
 
 #define LINUX_EEPROM_SIZE (E2END + 1)
