+++ Marlin/src/HAL/LPC1768/eeprom_api.h
@@ -1,10 +1,9 @@
 /**
  * Marlin 3D Printer Firmware
- *
  * Copyright (c) 2020 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
- * Copyright (c) 2016 Bob Cousins bobcousins42@googlemail.com
- * Copyright (c) 2015-2016 Nico Tonnhofer wurstnase.reprap@gmail.com
- * Copyright (c) 2016 Victor Perez victor_pv@hotmail.com
+ *
+ * Based on Sprinter and grbl.
+ * Copyright (c) 2011 Camiel Gubbels / Erik van der Zalm
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,11 +19,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-#include "../../inc/MarlinConfigPre.h"
+#pragma once
 
-#if EITHER(EEPROM_SETTINGS, SD_FIRMWARE_UPDATE)
+#include "../shared/eeprom_api.h"
 
-  #include "persistent_store_api.h"
-  PersistentStore persistentStore;
-
-#endif
+#define FLASH_EEPROM_EMULATION
