+++ Marlin/src/HAL/LPC1768/inc/Conditionals_post.h
@@ -21,6 +21,6 @@
  */
 #pragma once
 
-#if USE_EMULATED_EEPROM && NONE(SDCARD_EEPROM_EMULATION, SRAM_EEPROM_EMULATION)
+#if USE_FALLBACK_EEPROM && NONE(SDCARD_EEPROM_EMULATION, SRAM_EEPROM_EMULATION)
   #define FLASH_EEPROM_EMULATION
 #endif
