+++ Marlin/src/HAL/SAMD51/inc/Conditionals_post.h
@@ -21,8 +21,6 @@
  */
 #pragma once
 
-#if USE_EMULATED_EEPROM
-  #undef SRAM_EEPROM_EMULATION
-  #undef SDCARD_EEPROM_EMULATION
+#if USE_FALLBACK_EEPROM && NONE(SDCARD_EEPROM_EMULATION, SRAM_EEPROM_EMULATION)
   #define FLASH_EEPROM_EMULATION
 #endif
