+++ Marlin/src/HAL/STM32/inc/Conditionals_post.h
@@ -22,6 +22,6 @@
 #pragma once
 
 // If no real EEPROM, Flash emulation, or SRAM emulation is available fall back to SD emulation
-#if ENABLED(EEPROM_SETTINGS) && NONE(USE_REAL_EEPROM, FLASH_EEPROM_EMULATION, SRAM_EEPROM_EMULATION)
+#if ENABLED(EEPROM_SETTINGS) && NONE(USE_WIRED_EEPROM, FLASH_EEPROM_EMULATION, SRAM_EEPROM_EMULATION)
   #define SDCARD_EEPROM_EMULATION
 #endif
