+++ Marlin/src/HAL/STM32_F4_F7/eeprom.cpp
@@ -2,6 +2,9 @@
  * Marlin 3D Printer Firmware
  *
  * Copyright (c) 2020 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
+ * Copyright (c) 2016 Bob Cousins bobcousins42@googlemail.com
+ * Copyright (c) 2015-2016 Nico Tonnhofer wurstnase.reprap@gmail.com
+ * Copyright (c) 2016 Victor Perez victor_pv@hotmail.com
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,26 +21,15 @@
  *
  */
 
-#ifdef __STM32F1__
+#if defined(STM32GENERIC) && (defined(STM32F4) || defined(STM32F7))
 
-#include "../../inc/MarlinConfig.h"
+#include "../../inc/MarlinConfigPre.h"
 
-#if USE_REAL_EEPROM
+#if ENABLED(EEPROM_SETTINGS)
 
-#include "../shared/persistent_store_api.h"
+#include "../shared/eeprom_api.h"
 
-bool PersistentStore::access_start() {
-  #if ENABLED(SPI_EEPROM)
-    #if SPI_CHAN_EEPROM1 == 1
-      SET_OUTPUT(BOARD_SPI1_SCK_PIN);
-      SET_OUTPUT(BOARD_SPI1_MOSI_PIN);
-      SET_INPUT(BOARD_SPI1_MISO_PIN);
-      SET_OUTPUT(SPI_EEPROM1_CS);
-    #endif
-    spiInit(0);
-  #endif
-  return true;
-}
+bool PersistentStore::access_start() { return true; }
 bool PersistentStore::access_finish() { return true; }
 
 bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, uint16_t *crc) {
@@ -63,7 +55,7 @@ bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, ui
 bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t *crc, const bool writing/*=true*/) {
   do {
     uint8_t c = eeprom_read_byte((uint8_t*)pos);
-    if (writing && value) *value = c;
+    if (writing) *value = c;
     crc16(crc, &c, 1);
     pos++;
     value++;
@@ -73,5 +65,5 @@ bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t 
 
 size_t PersistentStore::capacity() { return E2END + 1; }
 
-#endif // USE_REAL_EEPROM
-#endif // __STM32F1__
+#endif // EEPROM_SETTINGS
+#endif // STM32GENERIC && (STM32F4 || STM32F7)
