+++ Marlin/src/HAL/STM32_F4_F7/inc/Conditionals_post.h
@@ -22,7 +22,7 @@
 #pragma once
 
 #if ENABLED(EEPROM_SETTINGS) && defined(STM32F7)
-  #undef USE_REAL_EEPROM
+  #undef USE_WIRED_EEPROM
   #undef SRAM_EEPROM_EMULATION
   #undef SDCARD_EEPROM_EMULATION
   #define FLASH_EEPROM_EMULATION
