+++ Marlin/src/pins/samd/pins_RAMPS_144.h
@@ -44,8 +44,9 @@
 //
 // EEPROM
 //
-#define E2END 0x7FFF                              // 32Kb (24lc256)
+//#define QSPI_EEPROM                             // Use AGCM4 onboard QSPI EEPROM (Uses 4K of RAM)
 #define I2C_EEPROM                                // EEPROM on I2C-0
+#define E2END 0x7FFF                              // 32K (24lc256)
 
 //
 // Limit Switches
