+++ platformio.ini
@@ -792,6 +792,7 @@ build_unflags  = -std=gnu++11
 src_filter     = ${common.default_src_filter} +<src/HAL/SAMD51>
 lib_deps       = ${common.lib_deps}
   SoftwareSerialM=https://github.com/FYSETC/SoftwareSerialM/archive/master.zip
+  Adafruit_SPIFlash=https://github.com/adafruit/Adafruit_SPIFlash/archive/master.zip
 debug_tool     = jlink
 
 #
